Split up CPU and GPU code path.
* Split up the code path for DMatrix initialization.
* Remove adapter explicit instantiation.
