+++ src/data/data.cu
@@ -163,20 +163,4 @@ void MetaInfo::SetInfo(const char * c_key, std::string const& interface_str) {
     LOG(FATAL) << "Unknown metainfo: " << key;
   }
 }
-
-template <typename AdapterT>
-DMatrix* DMatrix::Create(AdapterT* adapter, float missing, int nthread,
-                         const std::string& cache_prefix, size_t page_size) {
-  CHECK_EQ(cache_prefix.size(), 0)
-      << "Device memory construction is not currently supported with external "
-         "memory.";
-  return new data::SimpleDMatrix(adapter, missing, nthread);
-}
-
-template DMatrix* DMatrix::Create<data::CudfAdapter>(
-    data::CudfAdapter* adapter, float missing, int nthread,
-    const std::string& cache_prefix, size_t page_size);
-template DMatrix* DMatrix::Create<data::CupyAdapter>(
-    data::CupyAdapter* adapter, float missing, int nthread,
-    const std::string& cache_prefix, size_t page_size);
 }  // namespace xgboost
