+++ tests/cpp/helpers.cc
@@ -344,7 +344,7 @@ RandomDataGenerator::GenerateDMatrix(bool with_label, bool float_label,
 std::shared_ptr<DMatrix>
 GetDMatrixFromData(const std::vector<float> &x, int num_rows, int num_columns){
   data::DenseAdapter adapter(x.data(), num_rows, num_columns);
-  return std::shared_ptr<DMatrix>(new data::SimpleDMatrix(
+  return std::shared_ptr<DMatrix>(data::SimpleDMatrix::FromCPUData(
       &adapter, std::numeric_limits<float>::quiet_NaN(), 1));
 }
